ufile_transfer/requirement.txt
requests
mysqlclient
pyyaml
google-cloud-storage
pytest
click
python-magic